+++ spec/codec/pam_spec.rb
@@ -3,10 +3,10 @@ require 'spec_helper'
 RSpec.describe ImageUtil::Codec::Pam do
   it 'encodes and decodes a PAM image' do
     img = ImageUtil::Image.new(2, 1) { |x,_| ImageUtil::Color[x,0,0] }
-    data = described_class.encode(:pam, img, fill_to: 2)
+    data = described_class.encode(:pam, img)
     data.lines.first.chomp.should == 'P7'
     decoded = described_class.decode(:pam, data)
-    decoded.dimensions.should == [2,2]
+    decoded.dimensions.should == [2,1]
     decoded[1,0].should == ImageUtil::Color[1,0,0,255]
   end