+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - Redimension filter to change image dimensions
 - Redimension attempts to resize the existing buffer when increasing height or
   adding a new dimension
+- Sixel and Kitty codecs accept 1D images
+- `Pam.encode` no longer accepts `fill_to`; Sixel codecs pad images using the
+  redimension filter
 - Benchmarking subsystem with `bin/benchmark` using `benchmark-ips`
 - Improved heuristic for buffer resizing so multi-dimensional images stay intact